requirements.txt
graia-application-mirai
toml
apscheduler